+++ docs/CHANGELOG.md
@@ -8,15 +8,15 @@
 - added modal window to create new entries and append them to selectbox
 - added modal window to show entries short in popup
 - improved sql for similarity-search
-- added groupby trip for track-list
-- backend: fixed solr-import
+- added groupby trip for track-list, region for routes
+- fixed solr-import
 
 ### new features
 - backend/frontend: added statistics-section
 - backend/frontend: added linked route-attributes for track-routes
 - frontend: added modal window to create new entries and append them to selectbox
 - frontend: added modal window to show entries short in popup
-- frontend: added groupby trip for track-list
+- backend/frontend: added groupby trip for track-list, region for routes
 
 ### improvements
 - backend: added sort by name